+++ docs/changelog.md
@@ -29,6 +29,10 @@
 
 ## 更新说明
 
+1.6.80 fix: fix index_investing_global interface
+
+    1. 修复 index_investing_global 接口，同时修改该函数的参数
+
 1.6.79 fix: fix fund_rating_all interface
 
     1. 修复 fund_rating_all 接口，字段与目标网站对齐
@@ -1019,6 +1023,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.6.80 fix: fix index_investing_global interface
 
 1.6.79 fix: fix fund_rating_all interface